fix(linter/plugins): fix TS types for `RuleTester` (#16546)
Fix TS types for `RuleTester`. Merging the `RuleTester` class and namespace was working OK within the repo, but not in the built bundle. Fix that by defining the namespace in same file as the class.
